handle_read missing some "true" returns
Closes #52

When the interpreter gets unblocked, the handle_read function must return true. There were a few conditions where that was not happening, which would create a race condition in case there's data waiting in the buffer queue but where there isn't an extra notification.

I also added  a few extra debugging lines, which were necessary to help with this race condition.